+++ .copier/.copier-answers.yaml
@@ -1,5 +1,5 @@
 _commit: v1.2.1
 _src_path: https://github.com/armand-sauzay/python-template.git
-module_name: pybox
+module_name: izipyzi
 module_version: 0.1.0
-project_name: pybox
+project_name: izipyzi
